+++ packages/goods-ui/__tests__/goods-ui.test.js
@@ -1,7 +1,0 @@
-'use strict'
-
-const goodsUi = require('..')
-
-describe('goods-ui', () => {
-  it('needs tests')
-})