+++ packages/goods-ui/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 1.1.0 (2020-09-16)
+
+
+### Bug Fixes
+
+* **layout "w, h" boolean:** layout possible w and h using boolean will be '100%' ([e5bc470](https://github.com/PT-Tennova-Cipta-Inatech/pomona3-ui/commit/e5bc47077ab88a192101f979195286460b5973b1))
+
+
+### Features
+
+* **button:** add component button and setup goods-UI ([4ae09f1](https://github.com/PT-Tennova-Cipta-Inatech/pomona3-ui/commit/4ae09f13ee61da2810c7b8b73eaf85aa3de09cb8))
+* **button:** add component button and setup goods-UI ([4fba89a](https://github.com/PT-Tennova-Cipta-Inatech/pomona3-ui/commit/4fba89af9c03034e74243d89229c4c841d6a4a91))
 ## 1.0.1 (2020-09-16)
 
 **Note:** Version bump only for package goods-ui